refactor: split fulfillment.py (2465 lines) into queue + shipping modules (#229)
Split the largest endpoint file into two domain-focused modules:
- fulfillment_queue.py (1310 lines): production queue, QC, bulk updates
- fulfillment_shipping.py (1203 lines): shipping rates, labels, ship-from-stock

All URL paths remain identical (both use prefix="/fulfillment").
231 fulfillment tests pass unchanged.
